Add support for partner cloud providers configured in git repo (#11)
* Add support for partner cloud providers configured in git repo

The secret is created without a pre-install hook (which would force it to be installed before other objects).  According to issue https://github.com/helm/helm/issues/960 this annotation should not be required to ensure the secret gets created before the pod that requires it -- k8s should be smart enough to schedule things properly because the pod mounts the secret as a volume.